Another re-factoring round and removal of rangetype.py.
Apart form the eoxs_rangetype_list and eoxs_rangetype_load there was no other
EoxServer code using the range-type from/to dictionary conversion.
